cleaning(core): move paths
