+++ packages/core/src/filesystem/index.ts
@@ -1,3 +1,2 @@
-export * from "./paths";
 export * from "./pkg";
 export * from "./PrismicSharedConfig";
