+++ packages/core/src/node-utils/index.ts
@@ -1,2 +1,3 @@
 export * as Framework from "./framework";
 export * from "./manifest";
+export * from "./paths";
