+++ packages/core/src/node-utils/paths.ts
