+++ packages/init/src/steps/display-final-message.ts
@@ -1,7 +1,7 @@
-import { Utils, FileSystem, CONSTS } from "@slicemachine/core";
+import { Utils, CONSTS, NodeUtils } from "@slicemachine/core";
 
 export function displayFinalMessage(cwd: string): void {
-  const yarnLock = Utils.Files.exists(FileSystem.YarnLockPath(cwd));
+  const yarnLock = Utils.Files.exists(NodeUtils.YarnLockPath(cwd));
   const command = `${yarnLock ? "yarn" : "npm"} run ${CONSTS.SCRIPT_NAME}`;
 
   console.log(
