+++ packages/slice-machine/server/src/api/screenshots/generate.ts
@@ -1,7 +1,7 @@
 import type Models from "@slicemachine/core/build/models";
 import Files from "@lib/utils/files";
 import { BackendEnvironment } from "@lib/models/common/Environment";
-import { FileSystem } from "@slicemachine/core";
+import { NodeUtils } from "@slicemachine/core";
 import Puppeteer from "./puppeteer";
 import { resolvePathsToScreenshot } from "@slicemachine/core/build/libraries/screenshot";
 import {
@@ -49,10 +49,7 @@ export async function generateScreenshot(
 ): Promise<ScreenshotResults> {
   // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
   const slice: Models.SliceAsObject = Files.readJson(
-    FileSystem.CustomPaths(env.cwd)
-      .library(libraryName)
-      .slice(sliceName)
-      .model()
+    NodeUtils.CustomPaths(env.cwd).library(libraryName).slice(sliceName).model()
   );
 
   const variationIds: Models.VariationAsObject["id"][] = slice.variations.map(
@@ -109,7 +106,7 @@ async function generateForVariation(
     slice.id
   )}&vid=${encodeURIComponent(variationId)}`;
 
-  const pathToFile = FileSystem.GeneratedPaths(env.cwd)
+  const pathToFile = NodeUtils.GeneratedPaths(env.cwd)
     .library(libraryName)
     .slice(slice.name)
     .variation(variationId)
